# gym-app-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RankD API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/logged-sets": {
      "post": {
        "tags": [
          "logged-sets"
        ],
        "summary": "Create Logged Set",
        "operationId": "create_logged_set_logged_sets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoggedSetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedSetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logged-sets/bulk": {
      "post": {
        "tags": [
          "logged-sets"
        ],
        "summary": "Create Logged Sets Bulk",
        "operationId": "create_logged_sets_bulk_logged_sets_bulk_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoggedSetBulkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoggedSetOut"
                  },
                  "title": "Response Create Logged Sets Bulk Logged Sets Bulk Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logged-sets/me": {
      "get": {
        "tags": [
          "logged-sets"
        ],
        "summary": "List My Logged Sets",
        "operationId": "list_my_logged_sets_logged_sets_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gym Id"
            }
          },
          {
            "name": "exercise_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise Id"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoggedSetOut"
                  },
                  "title": "Response List My Logged Sets Logged Sets Me Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoggedSetBulkCreate": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LoggedSetCreate"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LoggedSetBulkCreate"
      },
      "LoggedSetCreate": {
        "properties": {
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "weight_kg": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Weight Kg"
          },
          "reps": {
            "type": "integer",
            "maximum": 50.0,
            "exclusiveMinimum": 0.0,
            "title": "Reps"
          },
          "added_weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 500.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Added Weight Kg"
          },
          "client_request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Request Id"
          }
        },
        "type": "object",
        "required": [
          "gym_id",
          "exercise_id",
          "weight_kg",
          "reps",
          "client_request_id"
        ],
        "title": "LoggedSetCreate"
      },
      "LoggedSetOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg"
          },
          "reps": {
            "type": "integer",
            "title": "Reps"
          },
          "added_weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Added Weight Kg"
          },
          "bodyweight_at_log_kg": {
            "type": "number",
            "title": "Bodyweight At Log Kg"
          },
          "estimated_1rm_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated 1Rm Kg"
          },
          "relative_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Ratio"
          },
          "points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "logged_at": {
            "type": "string",
            "format": "date-time",
            "title": "Logged At"
          },
          "client_request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Request Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "gym_id",
          "exercise_id",
          "weight_kg",
          "reps",
          "added_weight_kg",
          "bodyweight_at_log_kg",
          "estimated_1rm_kg",
          "relative_ratio",
          "points",
          "verified",
          "logged_at",
          "client_request_id"
        ],
        "title": "LoggedSetOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}